+++ deploy_api.py
@@ -52,6 +52,8 @@ async def idphoto_inference(
             size=size,
             head_measure_ratio=head_measure_ratio,
             head_height_ratio=head_height_ratio,
+            top_distance_max=top_distance_max,
+            top_distance_min=top_distance_min,
         )
     except FaceError:
         result_message = {"status": False}
